refactor(components): rename files, simplify structure
