+++ components/src/offerwall/components/index.ts
@@ -1,3 +1,3 @@
-export { AllSet } from './all-set/all-set.js'
-export { ContributionRequired } from './contribution-required/contribution-required.js'
-export { InstallRequired } from './install-required/install-required.js'
+export { AllSet } from './all-set.js'
+export { ContributionRequired } from './contribution-required.js'
+export { InstallRequired } from './install-required.js'
