+++ components/src/offerwall/components/install-required.css
